fix OAuth controller and refactor shibboleth config to pull attributes that are actually there
